backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
supabase>=2.0.0
httpx>=0.25.0
slowapi>=0.1.9
jinja2>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
